requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
certifi==2025.1.31
distro==1.9.0
dotenv==0.9.9
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.9.0
ollama==0.1.8
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.0
PyYAML==6.0.1
rsa==4.9
sniffio==1.3.1
Telethon==1.33.1
typing_extensions==4.12.2
